+++ android/src/main/kotlin/io/maido/intercom/IntercomFlutterPlugin.kt
@@ -1,6 +1,9 @@
 package io.maido.intercom
 
 import android.app.Application
+import androidx.annotation.NonNull
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
@@ -8,23 +11,41 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.PluginRegistry.Registrar
 import io.intercom.android.sdk.Company
 import io.intercom.android.sdk.Intercom
+import io.intercom.android.sdk.UnreadConversationCountListener
 import io.intercom.android.sdk.UserAttributes
 import io.intercom.android.sdk.identity.Registration
 import io.intercom.android.sdk.push.IntercomPushClient
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 
-
-
-class IntercomFlutterPlugin(private val application: Application) : MethodCallHandler {
+class IntercomFlutterPlugin : FlutterPlugin, MethodCallHandler, EventChannel.StreamHandler, ActivityAware {
   companion object {
+    @JvmStatic lateinit var application: Application
+
     @JvmStatic
     fun registerWith(registrar: Registrar) {
       val channel = MethodChannel(registrar.messenger(), "maido.io/intercom")
-      channel.setMethodCallHandler(IntercomFlutterPlugin(registrar.context() as Application))
-
+      application = registrar.context() as Application
+      channel.setMethodCallHandler(IntercomFlutterPlugin())
+      val unreadEventChannel = EventChannel(registrar.messenger(), "maido.io/intercom/unread")
+      unreadEventChannel.setStreamHandler(IntercomFlutterPlugin())
     }
   }
 
   private val intercomPushClient = IntercomPushClient()
+  private var unreadConversationCountListener: UnreadConversationCountListener? = null
+
+  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    val channel = MethodChannel(flutterPluginBinding.getFlutterEngine().getDartExecutor(), "maido.io/intercom")
+    channel.setMethodCallHandler(IntercomFlutterPlugin())
+    val unreadEventChannel = EventChannel(flutterPluginBinding.getFlutterEngine().getDartExecutor(), "maido.io/intercom/unread")
+    unreadEventChannel.setStreamHandler(IntercomFlutterPlugin())
+  }
+
+  // https://stackoverflow.com/a/62206235
+  override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+    application = binding.activity.getApplication()
+  }
 
   override fun onMethodCall(call: MethodCall, result: Result) {
     when {
@@ -91,8 +112,9 @@ class IntercomFlutterPlugin(private val application: Application) : MethodCallHa
         if(visibility != null) {
           Intercom.client().setInAppMessageVisibility(Intercom.Visibility.valueOf(visibility))
           result.success("Showing in app messages: $visibility")
+        } else {
+          result.success("Launched")
         }
-        result.success("Launched")
       }
       call.method == "unreadConversationCount" -> {
         val count = Intercom.client().unreadConversationCount
@@ -172,5 +194,33 @@ class IntercomFlutterPlugin(private val application: Application) : MethodCallHa
       }
       else -> result.notImplemented()
     }
+  }
+
+  override fun onListen(arguments: Any?, events: EventChannel.EventSink?) {
+    unreadConversationCountListener = UnreadConversationCountListener { count -> events?.success(count) }
+        .also {
+          Intercom.client().addUnreadConversationCountListener(it)
+        }
+  }
+
+  override fun onCancel(arguments: Any?) {
+    if (unreadConversationCountListener != null) {
+      Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    }
+  }
+
+  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+    if (unreadConversationCountListener != null) {
+      Intercom.client().removeUnreadConversationCountListener(unreadConversationCountListener)
+    }
+  }
+
+  override fun onDetachedFromActivity() {
+  }
+
+  override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+  }
+
+  override fun onDetachedFromActivityForConfigChanges() {
   }
 }