+++ test/intercom_flutter_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intercom_flutter/intercom_flutter.dart';
 
@@ -29,6 +30,13 @@ void main() {
       });
     });
 
+    test('testSendingAPNTokenToIntercom', () {
+      Intercom.sendTokenToIntercom('mock_apn_token');
+      expectMethodCall('sendTokenToIntercom', arguments: {
+        'token': 'mock_apn_token',
+      });
+    });
+
     group('registerIdentifiedUser', () {
       test('with userId', () {
         Intercom.registerIdentifiedUser(userId: 'test');
@@ -47,9 +55,9 @@ void main() {
       test('with userId and email should fail', () {
         expect(
           () => Intercom.registerIdentifiedUser(
-                userId: 'testId',
-                email: 'testEmail',
-              ),
+            userId: 'testId',
+            email: 'testEmail',
+          ),
           throwsArgumentError,
         );
       });
@@ -100,9 +108,10 @@ void main() {
 
     test('displayMessageComposer', () {
       Intercom.displayMessageComposer("message");
-      expectMethodCall('displayMessageComposer', arguments: {
-        "message" : "message"
-      });
+      expectMethodCall(
+        'displayMessageComposer',
+        arguments: {"message": "message"},
+      );
     });
 
     group('setInAppMessagesVisibility', () {
@@ -176,6 +185,30 @@ void main() {
         'name': 'TEST',
         'metaData': {'string': 'A string', 'number': 10, 'bool': true},
       });
+    });
+  });
+
+  group('UnreadMessageCount', () {
+    const String channelName = 'maido.io/intercom/unread';
+    const MethodChannel channel = MethodChannel(channelName);
+    final int value = 9;
+
+    setUp(() {
+      channel.setMockMethodCallHandler((MethodCall methodCall) async {
+        ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+          channelName,
+          const StandardMethodCodec().encodeSuccessEnvelope(value),
+          (ByteData data) {},
+        );
+      });
+    });
+
+    tearDown(() {
+      channel.setMockMethodCallHandler(null);
+    });
+
+    test('testStream', () async {
+      expect(await Intercom.getUnreadStream().first, value);
     });
   });
 }